An Android-based Indoor Positioning System for a public library.
This project is developed for a library which allows visitors to find
the location of the book with the help of an Indoor Positioning System.
The project has been uploaded on Google Play but the application has
been supported and excluded by zero devices. Any help would be
appreciated to identify the problem.